# internal/handler/auth.go
package handler

import (
	"github.com/gin-gonic/gin"
)

// @Summary SignUp
// @Tags auth
// @Description create account
// @ID create-account
// @Accept json
// @Produce json
// @Param input body domain.User true "account info"
// @Success 200 {integer} integer 1
// @Router /auth/sign-up [post]
func (h *Handler) signUp(c *gin.Context) {

}

// @Summary SignIn
// @Tags auth
// @Description login account
// @ID login-account
// @Accept json
// @Produce json
// @Param input body domain.User true "account info"
// @Success 200 {integer} integer 1
// @Router /auth/sign-in [post]
func (h *Handler) signIn(c *gin.Context) {

}